requirements.txt
# Core dependencies
agno-ai>=0.1.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
redis>=5.0.0
python-dotenv>=1.0.0

# Agno dependencies
openai>=1.0.0
duckduckgo-search>=3.8.0
lancedb>=0.3.0
pypdf>=3.0.0

# Additional utilities
python-multipart>=0.0.6
httpx>=0.24.0